requirements.txt
#aloe-django==0.1.3
#celery==3.1.24
#django-celery==3.1.17
#elasticsearch-dsl==2.1.0
#elasticsearch==2.4.0
#python-logstash==0.4.6
#sphinxcontrib-spelling


Django==2.0.7
Sphinx==1.7.5
coverage==4.5.1
django-cors-headers==2.3.0
django-extensions==2.0.7
django-filter==1.1.0
django-filters==0.2.1
django-mysql==2.3.0
djangorestframework==3.8.2
djangorestframework-xml==1.3.0
djangorestframework-yaml==1.0.3
drf-nested-routers==0.90.2
factory-boy==2.11.1
flake8==3.5.0
gevent==1.3.4
gunicorn==19.9.0
mysqlclient==1.3.13
pep8==1.7.1
pyflakes==2.0.0
setproctitle==1.1.10
sphinx_intl==0.9.11
sphinx-rtd-theme==0.4.0
sphinxcontrib-httpdomain==1.7.0
sphinxcontrib-plantuml==0.11
sphinxcontrib-websupport==1.1.0
elasticsearch-dsl==6.2.1

vcrpy-unittest==0.1.6

social-auth-app-django==2.1.0
